+++ mod-useFullSoak.ts
@@ -1,0 +1,3 @@
+/** @module useFullSoak */
+
+export { useFullSoak } from "./src/useFullSoak.ts";
